FIX Allowing multiple choice votes
Currently, if a user selects more than one choice for a multiple choice
poll only the first choice is recorded.

This is because `PollChoice::addVote()` checks if the user has voted or
not and only records the vote if they haven't but then it marks them as
having voted and all subsequent choices will be ignored.

Now we only execute `Poll::markAsVoted()` after all choices are recorded