NEW Abstracting the vote storage layer
Added an abstraction layer to the voting mechanism so that people can
extend the module to use their own logic and storage engine for voters.

Now it is possible to add a vote handler that will only allow Members to
vote (and only vote once) rather than relying on cookies.

Simply change the `Poll.vote_handler_class` config setting to a class
that extends Vote_Backend and you can define any logic to store and
validate votes